+++ app/views/shared/footer/_full.html.erb
@@ -34,7 +34,7 @@
                           target: '_blank' %>
             </li>
             <li class="mb-4">
-              <a href="https://flowbite.com/docs/components" class="hover:underline" target="_blank">Flowbite Components</a>
+              <a href="https://flowbite.com/docs/getting-started/introduction/" class="hover:underline" target="_blank">Component Library</a>
             </li>
           </ul>
         </div>
