+++ index.d.ts
@@ -1,0 +1,7 @@
+declare global {
+  interface Turbo {
+    session: {
+      drive: boolean
+    }
+  }
+}
